+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - New `history.frozen` API allows temporarily freezing history collection.
  - New `lousy.widget.zoom` statusbar widget: shows current page zoom level.
  - Added `log` signal, emitted whenever a message is logged.
+ - New `widget.is_alive` property. Can be accessed even if the widget has been destroyed.
 
 ### Fixed